register finalized to helpers
